requirements.txt
mtcnn==0.1.0
flickrapi==2.4.0
requests==2.26.0
python-dotenv==0.18.0